Feature/backend security

+++ backend/routers/posts_router.js
@@ -8,17 +8,52 @@ import { reverseGeoSearch } from "../reverseGeosearch.js";
 import { emitNewPostMessage } from "../socket.js";
 
 export const postsRouter = Router();
-//postsRouter.use(validateAccessToken);
+postsRouter.use(validateAccessToken);
 
 const postFiles = multer({ dest: "/usr/src/app/posts" });
 
 // create a new post
 postsRouter.post("/", postFiles.array("files"), async (req, res) => {
   try {
+    if (req.auth.payload.sub !== req.body.userId) {
+      return res.status(403).json({ error: "Post creation unauthorized." });
+    }
+
     const coordinates = JSON.parse(req.body.coordinates);
-    const longitude = coordinates[0];
-    const latitude = coordinates[1];
+    const longitude = Number(coordinates[0]);
+    const latitude = Number(coordinates[1]);
+
+    if (!Array.isArray(coordinates)) {
+      return res
+        .status(422)
+        .json({ error: "Post creation failed. Invalid coordinates." });
+    } else if (coordinates.length !== 2) {
+      return res
+        .status(422)
+        .json({ error: "Post creation failed. Invalid coordinates." });
+    } else if (Number.isNaN(longitude) || Number.isNaN(latitude)) {
+      return res
+        .status(422)
+        .json({ error: "Post creation failed. Invalid coordinates." });
+    }
+
     const location = await reverseGeoSearch(longitude, latitude);
+
+    if (
+      ![
+        "lithosphere",
+        "hydrosphere",
+        "biosphere",
+        "atmosphere",
+        "weather",
+        "space",
+        "other",
+      ].includes(req.body.type)
+    ) {
+      return res
+        .status(422)
+        .json({ error: "Post creation failed. Invalid type." });
+    }
 
     const post = await Post.create({
       description: req.body.description,
@@ -36,7 +71,17 @@ postsRouter.post("/", postFiles.array("files"), async (req, res) => {
     let fileMetadatas = [];
 
     req.files.forEach((file) => {
-      fileMetadatas.push({ metadata: file, PostId: post.id });
+      if (
+        file.mimetype.startsWith("image/") ||
+        file.mimetype.startsWith("video/") ||
+        file.mimetype.startsWith("audio/")
+      ) {
+        fileMetadatas.push({ metadata: file, PostId: post.id });
+      } else {
+        return res
+          .status(422)
+          .json({ error: "Post creation failed. File type not supported." });
+      }
     });
 
     let fileIds = [];
@@ -79,13 +124,20 @@ postsRouter.post("/", postFiles.array("files"), async (req, res) => {
 
     return res.json({ post, fileIds });
   } catch (e) {
+    console.log(e);
     return res.status(422).json({ error: "Post creation failed." });
   }
 });
 
 // get a single post
 postsRouter.get("/:id", async (req, res) => {
-  const post = await Post.findByPk(+req.params.id);
+  const id = Number(req.params.id);
+
+  if (!id) {
+    return res.status(404).json({ error: "Post not found." });
+  }
+
+  const post = await Post.findByPk(id);
 
   if (!post) {
     return res.status(404).json({ error: "Post not found." });
@@ -102,7 +154,13 @@ postsRouter.get("/:id", async (req, res) => {
 
 // update a specific post
 postsRouter.patch("/:id", async (req, res) => {
-  const post = await Post.findByPk(req.params.id);
+  const id = Number(req.params.id);
+
+  if (!id) {
+    return res.status(404).json({ error: "Post not found." });
+  }
+
+  const post = await Post.findByPk(id);
   const description = req.body.description;
   const coordinates = req.body.coordinates
     ? JSON.parse(req.body.coordinates)
@@ -116,8 +174,24 @@ postsRouter.patch("/:id", async (req, res) => {
   }
 
   if (description) update.description = description;
-  if (type) update.type = type;
   if (tags) update.tags = tags;
+  if (type) {
+    if (
+      ![
+        "lithosphere",
+        "hydrosphere",
+        "biosphere",
+        "atmosphere",
+        "weather",
+        "space",
+        "other",
+      ].includes(req.body.type)
+    ) {
+      return res.status(422).json({ error: "Post creation failed." });
+    } else {
+      update.type = type;
+    }
+  }
 
   let location = {
     location: "",
@@ -170,16 +244,26 @@ postsRouter.patch("/:id", async (req, res) => {
 
 // delete a specific post and all the related tags and files
 postsRouter.delete("/:id", async (req, res) => {
-  const post = await Post.findByPk(req.params.id);
+  const id = Number(req.params.id);
+
+  if (!id) {
+    return res.status(404).json({ error: "Post not found." });
+  }
+
+  const post = await Post.findByPk(id);
 
   if (post) {
-    await File.destroy({
-      where: {
-        PostId: post.id,
-      },
-    });
-    await searchIndex.deleteDocument(post.id);
-    await post.destroy();
+    if (req.auth.payload.sub !== post.UserId) {
+      return res.status(403).json({ error: "Post deletion unauthorized." });
+    } else {
+      await File.destroy({
+        where: {
+          PostId: post.id,
+        },
+      });
+      await searchIndex.deleteDocument(post.id);
+      await post.destroy();
+    }
   } else {
     return res.status(404).json({ error: "Post not found" });
   }
